pyproject.toml
[tool.poetry]
name = "reveration-time"
version = "0.0.1"
description = "Description of the project"
authors = ["Juan <rodriguezjuan001@outlook.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.11.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
plotly = "^5.15.0"
pandas = "^2.0.3"
numpy = "^1.25.2"
dask = "^2023.8.0"
scikit-learn = "^1.3.0"
scipy = "^1.11.1"
ipykernel = "^6.25.1"
regex = "^2023.8.8"
torch = {version = "^2.0.1+cpu", source = "pytorch"}
torchvision = {version = "^0.15.2+cpu", source = "pytorch"}
torchaudio = {version = "^2.0.2+cpu", source = "pytorch"}
python-dotenv = "^1.0.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
